\name{addCuts}
\alias{addCuts}
\title{Add proficiency levels}
\description{
This function transforms a continuous variable vector into a categorical vector.
}
\usage{addCuts(valVec, cutVec)}
\arguments{
  \item{valVec}{
  A continuous variable vector to be transformed into a catigorical vector.
}
  \item{cutVec}{
   Named vector indicating the proficiency level cut scores in descending(!) order.
}
}

\value{
A data frames.
}

\author{
Karoline Sachse
}

\examples{
}
