\name{calculateLinkErrorComponent}
\alias{calculateLinkErrorComponent}
\title{Add proficiency levels}
\description{
This function returns different linking error estimates.
}
\usage{calculateLinkErrorComponent(intParsT1, intParsT2, natParsT1, natParsT2)}
\arguments{
  \item{intParsT1}{
  A data frame containing the international item parameter estimates at time 1. 
  The first column has to be an ID variable, the second column contains the estimates.
}
 \item{intParsT2}{
  A data frame containing the international item parameter estimates at time 2. 
  The first column has to be an ID variable, the second column contains the estimates.
}
  \item{natParsT1}{
   A data frame containing the national item parameter estimates at time 1. 
  The first column has to be an ID variable, the second column contains the estimates.
}
  \item{natParsT2}{
   A data frame containing the national item parameter estimates at time 2. 
  The first column has to be an ID variable, the second column contains the estimates.
}
}

\value{
A list.
}

\author{
Karoline Sachse
}

\examples{
}
