\name{trendSp}
\alias{trendSp}
\title{Sparsamer Mean Trend}
\description{
This function calculates mean trends and the respective standard errors.
}
\usage{trendSp(meanTab1, meanTab2, linkerror)}
\arguments{
  \item{meanTab1}{
   A data frame containing "group", "mean_est", and "mean_se" at time 1.
}
 \item{meanTab2}{
A data frame containing "group", "mean_est", and "mean_se" at time 2.
}
  \item{linkerror}{
Numeric. 
}
}

\value{
A data frame containing the transformed parameters.
}

\author{
Karoline Sachse
}

\examples{
}
