\name{reportIQBlv}
\alias{reportIQBlv}
\title{IQB table generator}
\description{
This function returns basic IQB tables.
}
\usage{reportIQBlv(path, chapter=c("4", "5a", "5b", "6"),domain="Englisch Leseverstehen")}
\arguments{
  \item{path}{
	A character string containing eatTrend output file path.
}
  \item{chapter}{
   Character.
}
  \item{domain}{
   Character.
}
}

\value{
A data frame.
}

\author{
Karoline Sachse
}

\examples{
}
